requirements.txt
asgiref==3.9.1
attrs==25.3.0
click==8.2.1
coverage==7.10.6
dj-database-url==2.3.0
Django==5.0.7
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
gunicorn==22.0.0
h11==0.16.0
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==25.0
pluggy==1.6.0
psycopg==3.2.10
psycopg-binary==3.2.10
PyJWT==2.10.1
pytest==8.3.3
pytest-cov==5.0.0
pytest-django==4.9.0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.27.1
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
uvicorn==0.30.6
